requirements.txt
tornado
futures
riak
motor
schematics
psycopg2
queries
numpy
scipy
pyzmq
Jinja2
pandas
arrow
pycurl
ujson
pylibmc
protobuf
pycrypto
msgpack-python